+++ testing/testing_zungqr_m.cpp
@@ -55,7 +55,7 @@ int main( int argc, char** argv )
             n = opts.nsize[i];
             k = opts.ksize[i];
             if ( m < n || n < k ) {
-                printf( "skipping m %d, n %d, k %d because m < n or n < k\n", m, n, k );
+                printf( "skipping m %d, n %d, k %d because m < n or n < k\n", (int) m, (int) n, (int) k );
                 continue;
             }
             
@@ -85,7 +85,7 @@ int main( int argc, char** argv )
             magma_zsetmatrix( m, n, hA, lda, dA, ldda );
             magma_zgeqrf_gpu( m, n, dA, ldda, tau, dT, &info );
             if ( info != 0 )
-                printf("magma_zgeqrf_gpu returned error %d\n", info);
+                printf("magma_zgeqrf_gpu returned error %d\n", (int) info);
             magma_zgetmatrix( m, n, dA, ldda, hR, lda );
             magma_zgetmatrix( nb, min_mn, dT, nb, hT, nb );
             
@@ -94,7 +94,7 @@ int main( int argc, char** argv )
             gpu_time = magma_wtime() - gpu_time;
             gpu_perf = gflops / gpu_time;
             if ( info != 0 )
-                printf("magma_zungqr_gpu returned error %d\n", info);
+                printf("magma_zungqr_gpu returned error %d\n", (int) info);
             
             /* =====================================================================
                Performs operation using LAPACK
@@ -104,14 +104,14 @@ int main( int argc, char** argv )
                 
                 lapackf77_zgeqrf( &m, &n, hA, &lda, tau, h_work, &lwork, &info );
                 if ( info != 0 )
-                    printf("lapackf77_zgeqrf returned error %d\n", info);
+                    printf("lapackf77_zgeqrf returned error %d\n", (int) info);
                 
                 cpu_time = magma_wtime();
                 lapackf77_zungqr( &m, &n, &k, hA, &lda, tau, h_work, &lwork, &info );
                 cpu_time = magma_wtime() - cpu_time;
                 cpu_perf = gflops / cpu_time;
                 if ( info != 0 )
-                    printf("lapackf77_zungqr returned error %d\n", info);
+                    printf("lapackf77_zungqr returned error %d\n", (int) info);
                 
                 // compute relative error |R|/|A| := |Q_magma - Q_lapack|/|A|
                 blasf77_zaxpy( &n2, &c_neg_one, hA, &ione, hR, &ione );
